Fixes #24156 - Port robottelo tests - bookmarks - 2
